% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cached_get_eurostat_data}
\alias{cached_get_eurostat_data}
\title{Cached function to memoize the retrieval of Eurostat data.}
\usage{
cached_get_eurostat_data(nuts)
}
\value{
A memoized version of `get_population_data`.
}
\description{
This function uses memoization to cache the results of `get_population_data`,
avoiding repeated downloads of the data.
}
\details{
The function `cached_get_eurostat_data` memoizes the `get_population_data`
function using `memoise::memoise`, ensuring that the data is retrieved
from Eurostat only once and subsequent calls retrieve the cached results.
}
\keyword{internal}
