% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing_ids}
\alias{check_missing_ids}
\title{Identify and IDs without available data}
\usage{
check_missing_ids(df1, df2, df3, IDs, time_period)
}
\arguments{
\item{df1}{A dataframe containing columns `reporter`and `time_period`}

\item{df2}{A dataframe containing columns `reporter`and `time_period`}

\item{df3}{A dataframe containing columns `reporter`and `time_period`}

\item{IDs}{A vector specifying the time periods to be selected.}

\item{time_period}{A vector specifying the time periods to be selected.}
}
\value{
A vector of IDs that have zero values
}
\description{
This internal function checks for time periods within the specified 
dataframe where the `total_available` values are zero. It prints a warning 
message for each ID that has zero values in the selected time periods, 
advising the user to select other time periods if they want to include 
these IDs in the analysis. The function then returns a vector of IDs that 
have zero values in the specified time periods.
}
\keyword{internal}
