% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datatrade_EU}
\alias{datatrade_EU}
\title{datatrade_EU data}
\format{
List of 3 dataframes:

\describe{
\itemize{
\item \code{extra_import} A data frame external import with 296 rows and 4 columns:

\tabular{ll}{
  \code{reporter} \tab Country IDs.\cr
  \tab \cr
  \code{partner} \tab Time period.\cr
  \tab \cr
  \code{time_period} \tab Time period.\cr
  \tab \cr
  \code{value} \tab External import from countries where the pest is present.\cr
}
}
\itemize{
\item \code{intra_trade} A data frame internal trade with 2664 rows and 4 columns:

\tabular{ll}{
  \code{reporter} \tab Country IDs.\cr
  \tab \cr
  \code{partner} \tab Time period.\cr
  \tab \cr
  \code{time_period} \tab Time period.\cr
  \tab \cr
  \code{value} \tab External import from countries where the pest is present.\cr
}
}
\itemize{
\item \code{internal_production} A data frame internal production with 74 rows 
and 3 columns:

\tabular{ll}{
  \code{reporter} \tab Country IDs.\cr
  \tab \cr
  \code{time_period} \tab Time period.\cr
  \tab \cr
  \code{value} \tab External import from countries where the pest is present.\cr
}
}
}
}
\usage{
datatrade_EU
}
\description{
NUTS codes (Nomenclature of territorial units for statistics) for the European 
countries, including European Union Member States, 
EFTA countries, and candidate countries and potential candidates.
}
\keyword{datasets}
