% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datatrade_NorthAm}
\alias{datatrade_NorthAm}
\title{Example Trade Data for Northern America}
\format{
A list of three data frames:
\describe{
\item{extra_import}{Data on imports from countries outside of Northern America.
A data frame with 100 rows and 4 columns:
\describe{
\item{reporter}{ID of the receiving country in a two-letter country code (e.g., "US", "CA").}
\item{partner}{ID of the exporting country, coded as "CNTR_1" to "CNTR_5" to represent specific
non-Northern American countries.}
\item{time_period}{Time period of trade, in three-month periods ("January-March",
"April-June", "July-September", "October-December").}
\item{value}{Quantity of the commodity imported.}
}}
\item{intra_trade}{Data on internal trade within Northern American countries.
A data frame with 80 rows and 4 columns:
\describe{
\item{reporter}{ID of the receiving country in a two-letter code (e.g., "US", "CA").}
\item{partner}{ID of the exporting country in a two-letter code (e.g., "US", "CA").}
\item{time_period}{Time period of trade, in three-month periods ("January-March",
"April-June", "July-September", "October-December").}
\item{value}{Quantity of the commodity imported.}
}}
\item{internal_production}{Data on the internal production of the commodity within Northern America.
A data frame with 20 rows and 3 columns:
\describe{
\item{reporter}{ID of the producing country in a two-letter code (e.g., "US", "CA").}
\item{time_period}{Time period of trade, in three-month periods ("January-March",
"April-June", "July-September", "October-December").}
\item{value}{Quantity of the commodity produced.}
}}
}
}
\usage{
datatrade_NorthAm
}
\description{
\code{datatrade_NorthAm} simulated trade data for a commodity within Northern America.
This dataset illustrates the trade and production flow of a commodity that is a potential
pathway for the entry of a pest into Northern American countries in a year expressed
in three-month periods.
}
\keyword{datasets}
