% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pathwayapp.R
\name{distribution_panel}
\alias{distribution_panel}
\title{Creates a panel of conditional input elements based on the distribution type.}
\usage{
distribution_panel(ns, id_dist)
}
\arguments{
\item{id_dist}{Identifier for the distribution type.}
}
\value{
A tagList object containing conditional input elements for the specified distribution.
}
\description{
This function generates a dynamic panel of conditional input elements
depending on the specified distribution type.
}
\keyword{internal}
