% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_intra}
\alias{missing_intra}
\title{Identify and fill missing intra-partner combinations in a data frame}
\usage{
missing_intra(data, IDs)
}
\arguments{
\item{data}{A data frame containing partner and reporter columns.}

\item{IDs}{A vector specifying the IDs used for generating combinations.}
}
\value{
A data frame with missing intra-partner combinations filled with zero values.
}
\description{
This internal function identifies missing intra-partner combinations in a given
data frame and fills them with zero values. It is designed to be used within other functions.
}
\keyword{internal}
