% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pathwayapp.R
\name{n_from_dist}
\alias{n_from_dist}
\title{Generates random numbers from a specified distribution.}
\usage{
n_from_dist(dist, input_pars, n_iter)
}
\arguments{
\item{dist}{Distribution type (e.g., 'beta', 'binom', 'norm').}

\item{input_pars}{Comma-separated string of distribution parameters.}

\item{n_iter}{Number of iterations to generate random numbers.}
}
\value{
A numeric vector of random numbers generated from the specified distribution.
}
\description{
This function generates random numbers from a specified distribution based on
the input parameters and number of iterations.
}
\keyword{internal}
