% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrade.R
\name{ntrade}
\alias{ntrade}
\title{Ntrade calculation}
\usage{
ntrade(
  trade_data,
  filter_IDs = NULL,
  filter_period = NULL,
  summarise_result = NULL
)
}
\arguments{
\item{trade_data}{An object of class `TradeData` that can be the output of \code{\link{trade_data}}.}

\item{filter_IDs}{A vector containing the country IDs to filter (identification codes 
of the countries of interest). By default, it is set to \code{NULL}, meaning all 
\code{reporter} countries in the data frames will be considered.}

\item{filter_period}{A vector specifying the time periods to filter, based on 
the \code{time_period} column. By default, it is set to \code{NULL}, meaning 
all time periods in the data frames will be considered.}

\item{summarise_result}{A character vector specifying functions to summarise the \eqn{N_{trade}} result
for the selected time periods (\code{filter_period}). It accepts the expressions \code{"mean"} for the mean, 
\code{"sd"} for the standard deviation, "median" for the median value and
\code{"quantile(p)"} where \code{p} is the probability for the quantiles to the given probabilities. See examples.}
}
\value{
A data frame with the quantity of commodity imported by each countrie of interest 
from countries or regions where the pest is present. The result is returned for each 
time period if \code{summarise_result} is not specified (default is NULL). 
If a summary function is specified, the result will be summarised accordingly.
}
\description{
This function calculates the quantity of potentially infested imported commodity
(\eqn{N_{trade}}) from third countries where the pest is present.
}
\details{
The calculation of \eqn{N_{trade_i}} for each country of interest \eqn{i}
is based on the equation:
\deqn{N_{trade_i} = ExtraPest_i - ExtraPest_i \sum \frac{IntraExp_{ij}}{Total_i}
+ \sum ExtraPest_j \frac{IntraExp_{ji}}{Total_j},}
where:
\itemize{
\item \eqn{N_{trade_i}}: quantity of commodity from third countries remaining in
Member State \eqn{i} (\eqn{MS_i}), taking into account the direct importation from
a non-EU country where the pest is present, the re-exportation to other member states,
the indirect importation of the non-EU commodity from other member states.
\item \eqn{ExtraPest_i}: quantity of non-EU commodity imported (during the period
of time considered) by \eqn{MS_i} from countries where the pest is present (direct import).
\item \eqn{ExtraPest_j}: quantity of non-EU commodity imported (during the period
of time considered) by \eqn{MS_j} from countries where the pest is present (direct import).
\item \eqn{IntraExp_{ij}}: the quantity of commodity exported from \eqn{MS_i} to
\eqn{MS_j} (all origins confounded).
\item \eqn{IntraExp_{ji}}: the quantity of commodity exported from \eqn{MS_j} to
\eqn{MS_i} (all origins confounded).
\item \eqn{Total_i} or \eqn{Total_j}: total quantity of commodities available
in the \eqn{MS_i} or \eqn{MS_j}.
}
}
\examples{
## Example with simulated trade data for Northern America
data("datatrade_NA")
# Total extra-import data: data contains imports from 5 third countries (column partner). 
extra_total <- datatrade_NA$extra_import
# Extra-import data from countries where the pest is present (e.g., CNTR_1 and CNTR_2)
CNTR_pest <- c("CNTR_1", "CNTR_2")
extra_pest <- datatrade_NA$extra_import \%>\% filter(partner\%in\%CNTR_pest)
# Intra-trade data
intra_trade  <- datatrade_NA$intra_trade
# Internal production data
internal_production  <- datatrade_NA$internal_production
# Generate trade data (TradeData object)
trade_NA <- trade_data(extra_total = extra_total,
                       extra_pest = extra_pest,
                       intra_trade = intra_trade,
                       internal_production = internal_production)
# Calculation of the Ntrade for each time period
ntrade_NA <- ntrade(trade_data = trade_NA)
head(ntrade_NA)
# Ntrade summary for the time periods
ntrade_NA_summary <- ntrade(trade_data = trade_NA,
                            summarise_result = c("mean", "sd", 
                                                 "quantile(0.025)", 
                                                 "median",
                                                 "quantile(0.975)"))
head(ntrade_NA_summary)
# Plot the median of Ntrade
plot_countries(data = ntrade_NA_summary,
               IDs_column = "country_IDs", 
               values_column = "median") +
  xlim(-180,-20) + ylim(0,90)

## Example with simulated trade data for Europe 
# Load data
data("datatrade_EU")
# Total extra-import data: the total import is identified as partner "Extra_Total"
extra_total <- datatrade_EU$extra_import \%>\% filter(partner=="Extra_Total")
# Extra-import data from countries where the pest is present
extra_pest <- datatrade_EU$extra_import \%>\% filter(partner!="Extra_Total")
# Intra-trade data
intra_trade  <- datatrade_EU$intra_trade
# Internal production data
internal_production  <- datatrade_EU$internal_production
# Generate trade data (TradeData object)
trade_EU <- trade_data(extra_total = extra_total,
                       extra_pest = extra_pest,
                       intra_trade = intra_trade,
                       internal_production = internal_production)
# Ntrade mean and sd for the time periods
ntrade_EU <- ntrade(trade_data = trade_EU,
                    summarise_result = c("mean", "sd"))
# Plot Ntrade mean
plot_countries(data = ntrade_EU, 
               IDs_column="country_IDs", 
               values_column="mean") +
  xlim(-40,50) + ylim(25,70)
# Ntrade for selected countries and a specific time period
# Sample 5 countries from trade data
country_IDs <- sample(unique(trade_EU$total_trade$country_IDs), 5)
ntrade_EU_s <- ntrade(trade_data = trade_EU,
                      filter_IDs = country_IDs,
                      filter_period = 2020)
head(ntrade_EU_s)
# Plot Ntrade result
plot_countries(data = ntrade_EU_s, 
               IDs_column="country_IDs", 
               values_column="Ntrade_2020") +
  xlim(-40,50) + ylim(25,70)

}
