% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrade_app.R
\name{ntrade_app}
\alias{ntrade_app}
\title{Shiny app for Ntrade}
\usage{
ntrade_app()
}
\description{
Interactive application for the calculation and redistribution of the potentially
infected/infested quantity of commodities imported by a country from third countries
where the pest is present (\eqn{N_{trade}}) using the NUTS coding system. See
\href{https://ec.europa.eu/eurostat/web/nuts}{Nomenclature of territorial units for statistics}.
}
\examples{
\dontrun{
ntrade_app()
}
}
\seealso{
\code{\link[=ntrade]{ntrade()}}, \code{\link[=redist_nuts]{redist_nuts()}}, \code{\link[=redist_iso]{redist_iso()}}
}
