% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_app.R
\name{pathway_app}
\alias{pathway_app}
\title{Shiny app for pathway model}
\usage{
pathway_app()
}
\value{
No return value, called for side effects
}
\description{
Interactive application to estimate the number of potential founder populations (\eqn{NPFP})
of a pest in different regions or countries using the NUTS
coding system. See
\href{https://ec.europa.eu/eurostat/web/nuts}{Nomenclature of territorial units for statistics}.
}
\examples{
if (interactive()) {
pathway_app()
}
}
\seealso{
\code{\link[=pathway_model]{pathway_model()}}
}
