% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_model.R
\name{pathway_model}
\alias{pathway_model}
\title{Pathway model}
\usage{
pathway_model(
  ntrade_data,
  IDs_column,
  values_column,
  expression,
  parameters,
  niter = 100
)
}
\arguments{
\item{ntrade_data}{A data frame with the quantity of potentially infested commodities
imported from third countries where the pest is present. It can be calculated
using \code{\link{ntrade}} function.}

\item{IDs_column}{Column name in \code{ntrade_data} with the country IDs of interest.}

\item{values_column}{Column name in \code{ntrade_data} with the \eqn{N_{trade}} values
(quantity of potentially infested commodity imports) to be used in the pathway model.}

\item{expression}{Equation for the pathway model in the form of a string of characters.
This model must not include \eqn{N_{trade}}, since it is added multiplicatively to the
entered model by default. This equation is then added multiplicatively to:
\deqn{N_{inf_i} = N_{trade_i} \cdot \; ...}}

\item{parameters}{List with the names of the parameters included in \code{expression}
and the distribution of each one of them. The available distributions are:
\tabular{ll}{
"beta" \tab   see \code{\link[stats]{rbeta}}\cr
  \tab \cr
"binom" \tab  see \code{\link[stats]{rbinom}}\cr
  \tab \cr
"cauchy" \tab  see \code{\link[stats]{rcauchy}}\cr
  \tab \cr
"chisq" \tab  see \code{\link[stats]{rchisq}}\cr
  \tab \cr
"exp" \tab    see \code{\link[stats]{rexp}}\cr
  \tab \cr
"f"  \tab     see \code{\link[stats]{rf}}\cr
  \tab \cr
"gamma" \tab  see \code{\link[stats]{rgamma}}\cr
  \tab \cr
"geom" \tab   see \code{\link[stats]{rgeom}}\cr
  \tab \cr
"lnorm" \tab  see \code{\link[stats]{rlnorm}}\cr
  \tab \cr
"nbinom" \tab  see \code{\link[stats]{rnbinom}}\cr
  \tab \cr
"norm"  \tab  see \code{\link[stats]{rnorm}}\cr
  \tab \cr
"pois" \tab   see \code{\link[stats]{rpois}}\cr
  \tab \cr
"t"  \tab     see \code{\link[stats]{rt}}\cr
  \tab \cr
"unif" \tab   see \code{\link[stats]{runif}}\cr
  \tab \cr
"weibull" \tab see \code{\link[stats]{rweibull}}\cr
}}

\item{niter}{The number of iterations to generate random samples from the distributions.
Default 100 iterations.}
}
\value{
A dataframe with the statistics (mean, SD, minimum, first quartile, 
median, third quartile, and maximum) of founder population \eqn{N_{inf}} for each
country or region and for the total. [To be completed]
}
\description{
Estimate the amount of infested commodities that can enter into the
countries of interest imported from third countries where the pest is present,
and produce founder populations that can establish and spread.
}
\examples{
\dontrun{
# pathway model equation (is multiplied by Ntrade value)
eq <- "(1/p1)*p2*p3"
# distribution for each parameter
parameters <- list(
  p1 = list(dist = "beta", args = list(shape1 = 0.5, shape2 = 1)),
  p2 = list(dist = "gamma", args = list(shape = 1.5, scale = 100)),
  p3 = list(dist = "norm", args = list(mean = 5, sd = 2))
)
res <- pathway_model(ntrade_data = Nt_df,
                     IDs_column = "IDs",
                     values_column = "value",
                     expression = eq,
                     parameters = parameters,
                     niter = 100)
head(res)
# Total
res \%>\% filter(IDs == "Total")
}
}
