% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_countries.R
\name{plot_countries}
\alias{plot_countries}
\title{Plot trade values on a map}
\usage{
plot_countries(
  data,
  IDs_col,
  values_col,
  countrycode = "CNTR_CODE",
  colors = NULL,
  na_value = "grey",
  title = NULL,
  legend_title = NULL
)
}
\arguments{
\item{data}{A data frame containing the values to be plotted on the map.}

\item{IDs_col}{Column name in \code{data} with the country codes.
CNTR_ID from \code{\link[giscoR]{giscoR}} package .
[To be completed]}

\item{values_col}{Column name in \code{data} with the values to be plotted.}

\item{countrycode}{Country code in \code{IDs_col}. Possible values are "CNTR_CODE" 
(2-letter code of each country), or "ISO3_CODE" (3-letter code of each country). 
Default is "CNTR_CODE", for Europe the same as NUTS codes (country level), 
United Kingdom as "UK" and Greece as "EL". 
See \code{\link[giscoR]{gisco_countrycode}} for details on country codes.}

\item{colors}{Optional vector of colors used in the gradient scale.}

\item{na_value}{Color for missing values (default is "grey").}

\item{title}{A title for the plot (default is NULL).}

\item{legend_title}{A title for the legend. Default NULL, name in the \code{values_col}.}
}
\value{
A ggplot object with the plotted countries.
}
\description{
This function plots country values on a map using data provided and allows customization
of the aesthetics such as colors, legend title, and title.
}
\examples{
## Example with CNTR_CODE (2-letter country code)
# Country codes from the giscoR package
IDs  <- gisco_countrycode$CNTR_CODE[gisco_countrycode$continent=="Africa"]
# Values simulation
df <- data.frame(IDs = IDs,
                 value = runif(length(IDs)))
# Plot
plot_countries(data = df,
               IDs_col = "IDs",
               values_col = "value",
               countrycode = "CNTR_CODE")
# Changing colors and adding other ggplot2 options
pl <- plot_countries(data = df,
                     IDs_col = "IDs",
                     values_col = "value",
                     countrycode = "CNTR_CODE",
                     colors = c("white", "lightblue", "darkblue"))
pl + 
  xlim(-40, 60) + ylim(-40, 40) +
  theme_bw()

## Example with ISO3_CODE (3-letter country code)
# Country codes from the giscoR package
df_ISO3 <- data.frame(IDs = sample(gisco_countrycode$ISO3_CODE, 20),
                      value = runif(20))
plot_countries(data = df_ISO3,
               IDs_col = "IDs",
               values_col = "value",
               countrycode = "ISO3_CODE")

}
