% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nuts.R
\name{plot_nuts}
\alias{plot_nuts}
\title{Plot NUTS2 region values on a map}
\usage{
plot_nuts(
  data,
  nuts_col,
  values_col,
  nuts_level = 2,
  colors = NULL,
  na_value = "grey",
  title = NULL,
  legend_title = NULL
)
}
\arguments{
\item{data}{A data frame containing the values to be plotted on the map.}

\item{nuts_col}{Column name in \code{data} containing NUTS2 codes.}

\item{values_col}{Column name in \code{data} with the values to be plotted.}

\item{nuts_level}{Numeric. NUTS level to plot (0 = countries, 1 = groups of regions,
2 = regions, or 3 = provinces or districts). Default is 2.}

\item{colors}{Optional vector of colors used in the gradient scale.}

\item{na_value}{Color for missing values (default is "grey").}

\item{title}{A title for the plot (default is NULL).}

\item{legend_title}{A title for the legend. Default NULL, name in the \code{values_col}.}
}
\value{
A ggplot object with the plotted NUTS regions.
}
\description{
This function plots NUTS2 region values on a map using data provided and allows customization
of the aesthetics such as colors, legend title, and title.
}
\examples{
\dontrun{
plot_nuts(data = nt_redistrib, nuts_col = "NUTS2", values_col="mean")
}
}
