% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nuts2.R
\name{plot_nuts2}
\alias{plot_nuts2}
\title{Plot NUTS2 region values on a map}
\usage{
plot_nuts2(
  data,
  nuts2_column,
  values_column,
  colors = NULL,
  na.value = "grey",
  title = NULL,
  legend_title = NULL
)
}
\arguments{
\item{data}{A data frame containing the values to be plotted on the map.}

\item{nuts2_column}{Column name in \code{data} containing NUTS2 codes.}

\item{values_column}{Column name in \code{data} with the values to be plotted.}

\item{colors}{Optional vector of colors used in the gradient scale.}

\item{na.value}{Color for missing values (default is "grey").}

\item{title}{A title for the plot (default is NULL).}

\item{legend_title}{A title for the legend. Default NULL, name in the \code{values_column}.}
}
\value{
A ggplot object with the plotted NUTS regions.
}
\description{
This function plots NUTS2 region values on a map using data provided and allows customization
of the aesthetics such as colors, legend title, and title.
}
\examples{
\dontrun{
plot_nuts2(data = nt_redistrib, nuts2_column = "NUTS2", values_column="mean")
}
}
