% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pathwayapp.R
\name{q_from_dist}
\alias{q_from_dist}
\title{Generates quantiles from a specified distribution.}
\usage{
q_from_dist(dist, input_pars)
}
\arguments{
\item{dist}{Distribution type (e.g., 'beta', 'binom', 'norm').}

\item{input_pars}{Comma-separated string of distribution parameters.}
}
\value{
A numeric vector.
}
\description{
This function generates quantiles (0.05, 0.25, 0.5, 0.75, 0.95) from a 
specified distribution based on the input parameters.
}
\keyword{internal}
