% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_data}
\alias{summarise_data}
\title{Data preparation}
\usage{
summarise_data(data, select_period = NULL, reporter = TRUE, partner = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{select_period}{A vector specifying the time periods to be selected (from time_period column).
By default, it is set to NULL, meaning all periods in the data frames will be considered.}

\item{reporter}{Logical. If TRUE (default) values are grouped by reporter.}

\item{partner}{Logical. If TRUE values are grouped by partner (default is FALSE).}
}
\value{
A data frame with the values grouped by ID for each selected time period.
ID corresponds to reporter if TRUE or to partner if TRUE.
If reporter and partner are TRUE the data frame contains both columns.
}
\description{
Data preparation
}
\keyword{internal}
