% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data.R
\name{trade_data}
\alias{trade_data}
\title{Prepare Trade Databases}
\usage{
trade_data(
  extra_total,
  extra_pest,
  intra_trade,
  internal_production,
  filter_IDs = NULL,
  filter_period = NULL
)
}
\arguments{
\item{extra_total}{A data frame containing the total quantity of commodity imported
from third countries. It must contain the following columns: \code{reporter}, 
\code{partner}, \code{value} and \code{time_period}.}

\item{extra_pest}{A data frame containing the quantity of commodity imported by
from third countries where the pest is present. It must contain the following columns: 
\code{reporter}, \code{partner}, \code{value} and \code{time_period}.}

\item{intra_trade}{A data frame containing the quantity of commodity imported
and exported among the countries of interest.  It must contain the following columns: 
\code{reporter}, \code{partner}, \code{value} and \code{time_period}.}

\item{internal_production}{A data frame containing the quantity of commodity
produced by each country of interest. It must contain the following columns: 
\code{reporter}, \code{value} and \code{time_period}.}

\item{filter_IDs}{A vector containing the country IDs to filter (identification codes 
of the countries of interest). By default, it is set to \code{NULL}, meaning all 
\code{reporter} countries in the data frames will be considered.}

\item{filter_period}{A vector specifying the time periods to filter, based on 
the \code{time_period} column. By default, it is set to \code{NULL}, meaning 
all time periods in the data frames will be considered.}
}
\value{
An object of class \code{TradeData} is returned containing the following list of dataframes:

\itemize{
\item \code{total_trade} A data frame with one row for each ID and each time period with 9 variables:

\tabular{ll}{
  \code{country_IDs} \tab Country IDs.\cr
  \tab \cr
  \code{time_period} \tab Time period.\cr
  \tab \cr
  \code{extra_total} \tab Total external import.\cr
  \tab \cr
  \code{extra_pest} \tab External import from countries where the pest is present.\cr
  \tab \cr
  \code{intra_import} \tab Internal import between IDs.\cr
  \tab \cr
  \code{intra_export} \tab Internal export between IDs.\cr
  \tab \cr
  \code{internal_production} \tab Internal production.\cr
  \tab \cr
  \code{total_available} \tab Total available.\cr
  \tab \cr
  \code{export_prop} \tab Proportion of internal export to the total available.
  [0, 1) if \code{intra_export} > \code{total_available}, otherwise 
  1 (\code{intra_export} <= \code{total_available}).\cr
}
}
\itemize{
\item \code{intra_trade} A data frame with values of trade commodity between IDs:

\tabular{ll}{
  \code{reporter} \tab Country IDs importing.\cr
  \tab \cr
  \code{partner} \tab Country IDs exporting.\cr
  \tab \cr
  \code{time_period} \tab Time period.\cr
  \tab \cr
  \code{value} \tab Quantity of commodity.\cr
  \tab \cr
  \code{export_prop} \tab Proportion of internal export to total available according
  to the specified proportion for each partner.\cr
}
The quantity of exported commodity (\code{value}) is multiplied by \code{export_prop} of each
partner in order not to exceed the total available commodity in each ID.
}
}
\description{
This function prepares trade databases by providing the total available commodity
for each country of interest based on the provided data.
}
\details{
Trade flow from reporter to partner.
Total available = extra total + internal production.
Missing values as zeros.
Proportional internal export when intra export is higher than the total available.
}
\examples{
## Example with simulated trade data for Northern America
# Load data
data("datatrade_NA")
# Total extra-import data: data contains imports from 5 third countries (column partner). 
extra_total <- datatrade_NA$extra_import
# Extra-import data from countries where the pest is present (e.g., CNTR_1 and CNTR_2)
CNTR_pest <- c("CNTR_1", "CNTR_2")
extra_pest <- datatrade_NA$extra_import \%>\% filter(partner\%in\%CNTR_pest)
# Intra-trade data
intra_trade  <- datatrade_NA$intra_trade
# Internal production data
internal_production  <- datatrade_NA$internal_production
# Generate trade data (TradeData object)
trade_NA <- trade_data(extra_total = extra_total,
                       extra_pest = extra_pest,
                       intra_trade = intra_trade,
                       internal_production = internal_production)
head(trade_NA$total_trade)
head(trade_NA$intra_trade)
# Plot the total available quantity of commodity available in each country
plot_countries(data = trade_NA$total_trade,
               IDs_column = "IDs", 
               values_column = "total_available") +
  xlim(-180,-20) + ylim(0,90)
  
## Example with simulated trade data for Europe 
# with selected countries and a specific time period
# Load data
data("datatrade_EU")
# Total extra-import data: the total import is identified as partner "Extra_Total"
extra_total <- datatrade_EU$extra_import \%>\% filter(partner=="Extra_Total")
# Extra-import data from countries where the pest is present
extra_pest <- datatrade_EU$extra_import \%>\% filter(partner!="Extra_Total")
# Intra-trade data
intra_trade  <- datatrade_EU$intra_trade
# Internal production data
internal_production  <- datatrade_EU$internal_production
# Sample 5 countries from data
filter_IDs <- sample(unique(extra_total$reporter), 5)
# Generate trade data (TradeData object)
trade_EU <- trade_data(extra_total = extra_total,
                       extra_pest = extra_pest,
                       intra_trade = intra_trade,
                       internal_production = internal_production,
                       filter_IDs = filter_IDs,
                       filter_period = 2020)
# Plot the total available quantity of commodity available in each country
plot_countries(data = trade_EU$total_trade, 
               IDs_column = "IDs", 
               values_column = "total_available") +
  xlim(-40,50) + ylim(25,70)

  
}
