% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_population_data}
\alias{get_population_data}
\title{Population data}
\usage{
get_population_data(nuts)
}
\arguments{
\item{nuts}{Numeric, NUTS level to obtain population data}
}
\value{
A data frame with population data for each NUTS level
}
\description{
This internal function download Eurostat population data
}
\keyword{internal}
