% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrade.R
\name{ntrade}
\alias{ntrade}
\title{Ntrade calculation}
\usage{
ntrade(trade, IDs = NULL, select_period = NULL, summarize_ntrade = NULL)
}
\arguments{
\item{trade}{An object of class `TradeData` that can be the output of \code{\link{trade_data}}.}

\item{IDs}{A vector containing the identification codes of the countries of interest.}

\item{select_period}{A vector specifying the time periods to be selected (from the \code{time_period} column).
By default, it is set to NULL, meaning all periods in the trade data will be considered.}

\item{summarize_ntrade}{A character vector specifying functions to summarize the \eqn{N_{trade}} result
for the selected time periods (\code{select_period}). It accepts the expressions \code{"mean"} for the mean, 
\code{"sd"} for the standard deviation, "median" for the median value and
\code{"quantile(p)"} where \code{p} is the probability for the quantiles to the given probabilities. See examples.}
}
\value{
A data frame with the quantity of commodity imported by each ID from countries or regions
  where the pest is present. The result is returned for each time period if \code{summarize_ntrade} is not specified
  (default is NULL). If a summary function is specified, the result will be summarized accordingly.
}
\description{
This function calculates the quantity of potentially infested imported commodity
(\eqn{N_{trade}}) from third countries where the pest is present.
}
\details{
The calculation of \eqn{N_{trade_i}} for each country of interest \eqn{i}
is based on the equation:
\deqn{N_{trade_i} = ExtraPest_i - ExtraPest_i \sum \frac{IntraExp_{ij}}{Total_i}
+ \sum ExtraPest_j \frac{IntraExp_{ji}}{Total_j},}
where:
\itemize{
\item \eqn{N_{trade_i}}: quantity of commodity from third countries remaining in
Member State \eqn{i} (\eqn{MS_i}), taking into account the direct importation from
a non-EU country where the pest is present, the re-exportation to other member states,
the indirect importation of the non-EU commodity from other member states.
\item \eqn{ExtraPest_i}: quantity of non-EU commodity imported (during the period
of time considered) by \eqn{MS_i} from countries where the pest is present (direct import).
\item \eqn{ExtraPest_j}: quantity of non-EU commodity imported (during the period
of time considered) by \eqn{MS_j} from countries where the pest is present (direct import).
\item \eqn{IntraExp_{ij}}: the quantity of commodity exported from \eqn{MS_i} to
\eqn{MS_j} (all origins confounded).
\item \eqn{IntraExp_{ji}}: the quantity of commodity exported from \eqn{MS_j} to
\eqn{MS_i} (all origins confounded).
\item \eqn{Total_i} or \eqn{Total_j}: total quantity of commodities available
in the \eqn{MS_i} or \eqn{MS_j}.
}
}
\examples{
## Example with simulated trade data for Northern America
data("datatrade_NA")
library(dplyr)
# Total extra-import data: data contains imports from 5 third countries (column partner). 
extra_total <- datatrade_NA$extra_import
# Extra-import data from countries where the pest is present (e.g., CNTR_1 and CNTR_2)
CNTR_pest <- c("CNTR_1", "CNTR_2")
extra_pest <- datatrade_NA$extra_import \%>\% filter(partner\%in\%CNTR_pest)
# Intra-trade data
intra_trade  <- datatrade_NA$intra_trade
# Internal production data
internal_production  <- datatrade_NA$internal_production
# Generate trade data (TradeData object)
trade_NA <- trade_data(extra_total = extra_total,
                       extra_pest = extra_pest,
                       intra_trade = intra_trade,
                       internal_production = internal_production)
# Calculation of the Ntrade for each time period
ntrade_NA <- ntrade(trade = trade_NA)
head(ntrade_NA)
# Ntrade summary for the time periods
ntrade_NA_summary <- ntrade(trade = trade_NA,
                            summarize_ntrade = c("mean", "sd", 
                                                 "quantile(0.025)", 
                                                 "median",
                                                 "quantile(0.975)"))
head(ntrade_NA_summary)
# Plot the median of Ntrade
library(ggplot2)
plot_countries(data = ntrade_NA_summary,
               IDs_column = "IDs", 
               values_column = "median") +
  xlim(-180,-20) + ylim(0,90)

## Example with simulated trade data for Europe 
# Load data
data("datatrade_EU")
# Total extra-import data: the total import is identified as partner "Extra_Total"
extra_total <- datatrade_EU$extra_import \%>\% filter(partner=="Extra_Total")
# Extra-import data from countries where the pest is present
extra_pest <- datatrade_EU$extra_import \%>\% filter(partner!="Extra_Total")
# Intra-trade data
intra_trade  <- datatrade_EU$intra_trade
# Internal production data
internal_production  <- datatrade_EU$internal_production
# Generate trade data (TradeData object)
trade_EU <- trade_data(extra_total = extra_total,
                       extra_pest = extra_pest,
                       intra_trade = intra_trade,
                       internal_production = internal_production)
# Ntrade mean and sd for the time periods
ntrade_EU <- ntrade(trade = trade_EU,
                    summarize_ntrade = c("mean", "sd"))
# Plot Ntrade mean
plot_countries(data = ntrade_EU, 
               IDs_column="IDs", 
               values_column="mean") +
  xlim(-40,50) + ylim(25,70)
# Ntrade for selected countries and a specific time period
# Sample 5 countries from trade data
IDs <- sample(unique(trade_EU$total_trade$IDs), 5)
ntrade_EU_s <- ntrade(trade = trade_EU,
                      IDs = IDs,
                      select_period = 2020)
head(ntrade_EU_s)
# Plot Ntrade result
plot_countries(data = ntrade_EU_s, 
               IDs_column="IDs", 
               values_column="Ntrade_2020") +
  xlim(-40,50) + ylim(25,70)

}
