% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrade_redist.R
\name{ntrade_redist}
\alias{ntrade_redist}
\title{Ntrade redistribution}
\usage{
ntrade_redist(
  ntrade_data,
  ntrade_nuts_col,
  ntrade_values_col,
  to_nuts = 2,
  redist_data = "population",
  redist_nuts_col = NULL,
  redist_values_col = NULL,
  population_year = 2023
)
}
\arguments{
\item{ntrade_data}{A data frame with the quantity of potentially infested commodities
imported from third countries where the pest is present. It can be calculated
using \code{\link{ntrade}} function.}

\item{ntrade_nuts_col}{Column name in \code{ntrade_data} with the NUTS0 codes.}

\item{ntrade_values_col}{Column name, or vector with the name of multiple columns,
in \code{ntrade_data} with the \eqn{N_{trade}} values
(quantity of potentially infested commodity imports) to be redistributed.}

\item{to_nuts}{Numeric. NUTS level for redistribution (1, 2 or 3).}

\item{redist_data}{Data frame to proportionally distribute \eqn{N_{trade}}.
By default NULL: redistribution based on population
(Eurostat data - <https://ec.europa.eu/eurostat/databrowser/product/page/demo_r_pjangrp3>).}

\item{redist_nuts_col}{Column name in \code{redist_data} with the destination
NUTS codes in the redistribution. Same NUTS level as set in \code{to_nuts}.}

\item{redist_values_col}{Column name in \code{redist_data} with the with the values
according to which to proportionally redistribute \eqn{N_{trade}}.}

\item{population_year}{Year of population data. By default 2023.
Available years can be found at
<https://ec.europa.eu/eurostat/databrowser/product/page/demo_r_pjangrp3>.
Multiple years can be entered by a numeric vector, the average population of
the different years will be used.}
}
\value{
A data frame with the quantity of commodity redistributed.
}
\description{
Redistribution of the quantity of potentially infested commodities (\eqn{N_{trade}})
of each NUTS0 (country) among smaller territorial units (NUTS1, NUTS2 or NUTS3).
}
\details{
NUTS stands for nomenclature of territorial units for statistics.
\eqn{N_{trade}} can be generated using the \code{\link{ntrade}} function.
}
\examples{
\dontrun{
Nt_r <- ntrade_redist(
  ntrade_data = Nt_df,
  ntrade_nuts_col = "country_IDs",
  ntrade_values_col = "mean",
  to_nuts = 2
)
}
}
