% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrade_redist.R
\name{ntrade_redist}
\alias{ntrade_redist}
\title{Ntrade redistribution}
\usage{
ntrade_redist(
  ntrade_data,
  nuts_column,
  values_column,
  to_nuts = 2,
  prop_data = NULL,
  prop_nuts_column = NULL,
  prop_values_column = NULL,
  time_period = 2023
)
}
\arguments{
\item{ntrade_data}{A data frame with the quantity of potentially infested commodities
imported from third countries where the pest is present. It can be calculated
using \code{\link{ntrade}} function.}

\item{nuts_column}{Column name in \code{ntrade_data} with the NUTS0 codes.}

\item{values_column}{Column name, or vector with the name of multiple columns, 
in \code{ntrade_data} with the \eqn{N_{trade}} values
(quantity of potentially infested commodity imports) to be redistributed.}

\item{to_nuts}{Numeric. NUTS level for redistribution (1, 2 or 3).}

\item{prop_data}{Data frame to proportionally distribute \eqn{N_{trade}}.
By default NULL: redistribution based on population
(Eurostat data - <https://ec.europa.eu/eurostat/databrowser/product/page/demo_r_pjangrp3>).}

\item{prop_nuts_column}{Column name in \code{prop_data} with the destination
NUTS codes in the redistribution. Same NUTS level as set in \code{to_nuts}.}

\item{prop_values_column}{Column name in \code{prop_data} with the with the values
according to which to proportionally redistribute \eqn{N_{trade}}.}

\item{time_period}{Year of population data. By default 2023.
Available years can be found at
<https://ec.europa.eu/eurostat/databrowser/product/page/demo_r_pjangrp3>.
Multiple years can be entered by a numeric vector, the average population of
the different years will be used.}
}
\value{
A data frame with the quantity of commodity redistributed.
}
\description{
Redistribution of the quantity of potentially infested commodities (\eqn{N_{trade}})
of each NUTS0 (country) among smaller territorial units (NUTS1, NUTS2 or NUTS3).
}
\details{
NUTS stands for nomenclature of territorial units for statistics.
\eqn{N_{trade}} can be generated using the \code{\link{ntrade}} function.
}
\examples{
\dontrun{
Nt_r <- ntrade_redist(
ntrade_data = Nt_df,
nuts_column = "IDs",
values_column = "mean",
to_nuts = 2
)
}
}
