% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrade_redist.R, R/ntrade_redist_iso.R
\name{ntrade_redist}
\alias{ntrade_redist}
\title{Ntrade redistribution}
\usage{
ntrade_redist(
  ntrade_data,
  ntrade_iso_col,
  ntrade_values_col,
  redist_data,
  redist_iso_col,
  redist_values_col
)

ntrade_redist(
  ntrade_data,
  ntrade_iso_col,
  ntrade_values_col,
  redist_data,
  redist_iso_col,
  redist_values_col
)
}
\arguments{
\item{ntrade_data}{A data frame containing the quantity of potentially infested 
commodities imported from third countries where the pest is present. This data can 
be generated using \code{\link{ntrade}} function.}

\item{ntrade_iso_col}{A string specifying the column name in \code{ntrade_data} 
with the ISO 3166-1 codes.}

\item{ntrade_values_col}{A string or vector specifying the column name(s) in 
\code{ntrade_data} with the \eqn{N_{trade}} values (quantity of potentially 
infested commodity imports) to be redistributed.}

\item{redist_data}{A data frame to provide custom proportions for redistributing 
\eqn{N_{trade}}.}

\item{redist_iso_col}{A string specifying the column name in \code{redist_data} 
that contains the destination ISO 3166-2 codes.}

\item{redist_values_col}{A string specifying the column name in \code{redist_data} 
with the values for proportional redistribution. This will define the weights 
used for redistributing \eqn{N_{trade}}.}

\item{ntrade_nuts_col}{A string specifying the column name in \code{ntrade_data} 
with the NUTS0 codes.}

\item{to_nuts}{A numeric value (1, 2, or 3) specifying the NUTS level for redistribution.
Default 2, indicating redistribution to NUTS2.}

\item{redist_nuts_col}{A string specifying the column name in \code{redist_data} 
that contains the destination NUTS codes. The NUTS level should correspond to the 
value specified in  \code{to_nuts}. \code{NULL} (default) if a data frame is not 
incorporated in \code{redist_data}.}

\item{population_year}{A numeric value specifying the year of population data 
to use in the redistribution. only necessary if \code{"population"} is specified in 
\code{redist_data} (default is 2023). If multiple years are provided, the average 
population across those years will be used. Available years can be found at
\link[https://ec.europa.eu/eurostat/databrowser/product/page/demo_r_pjangrp3]{population 
data from Eurostat population data}.}
}
\value{
A data frame with the redistributed quantity of potentially infested 
commodities across the specified NUTS level.

A data frame with the redistributed quantity of potentially infested 
commodities across the specified subnational level.
}
\description{
Redistribution of the quantity of potentially infested commodities (\eqn{N_{trade}})
from country-level (NUTS0) data to smaller territores (NUTS1, NUTS2 or NUTS3).
See \link[https://ec.europa.eu/eurostat/web/nuts]{NUTS - Nomenclature of territorial 
units for statistics}.

Redistribution of the quantity of potentially infested commodities (\eqn{N_{trade}})
from country-level (ISO 3166-1) data to principal subdivisions (ISO 3166-2).
See \link[https://www.iso.org/iso-3166-country-codes.html]{ISO 3166 Maintenance Agency}.
}
\details{
This function enables redistribution of trade data from national-level NUTS0 
to smaller territorial units (NUTS1, NUTS2, or NUTS3), either proportionally 
based on population data from Eurostat or using user-supplied redistribution proportions. 
Population data for redistribution is automatically fetched for the specified time 
period from the Eurostat database.

This function enables redistribution of trade data from national-level  
to smaller territorial units (NUTS1, NUTS2, or NUTS3), proportionally 
to user-supplied redistribution proportions.
}
\examples{
\dontrun{
Nt_r <- ntrade_redist(
  ntrade_data = Nt_df,
  ntrade_nuts_col = "country_IDs",
  ntrade_values_col = "mean",
  to_nuts = 2
)
}
\dontrun{
Nt_r <- ntrade_redist(
  ntrade_data = Nt_df,
  ntrade_iso_col = "country_IDs",
  ntrade_values_col = "mean",
  redist_data = df_population,
  redist_iso_col = "ISO_3166_2",
  redist_values_col = "population"
)
}
}
