% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_app.R
\name{pathway_app}
\alias{pathway_app}
\title{Shiny app for pathway model}
\usage{
pathway_app()
}
\description{
Interactive application to estimate the number of founder populations (\eqn{NPFP})
of a pest in different regions or countries, for countries included in the NUTS
coding system. See
\href{https://ec.europa.eu/eurostat/web/nuts}{NUTS - Nomenclature of territorial units for statistics}.
}
\examples{
\dontrun{
pathway_app()
}
}
\seealso{
\code{\link[=pathway_model]{pathway_model()}}
}
