% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPRAentry-package.R
\docType{package}
\name{qPRAentry-package}
\alias{qPRAentry}
\alias{qPRAentry-package}
\title{qPRAentry: Tools to support the entry step of quantitative Pest Risk Assessment}
\description{
The \bold{R} package \bold{qPRAentry} is provided to support risk assessors in performing the entry step of the quantitative pest risk assessments. The method expresses pest entry providing a distribution of the number of potential founder populations that can establish in the risk assessment area as a result of entry. The package encompasses functions for calculating and redistributing the quantity of commodities imported from countries where a pest of interest is present. The package and functions used are based on the work of the European Food Safety Authority (EFSA PLH Panel et al., 2018) \url{https://doi.org/10.2903/J.EFSA.2018.5350}.
}
\author{
\strong{Maintainer}: Martina Cendoya \email{cendoya_marmar@gva.es} (\href{https://orcid.org/0000-0002-0134-7781}{ORCID})

Authors:
\itemize{
  \item Maria Chiara Rosace \email{mariachiara.rosace@unicatt.it} (\href{https://orcid.org/0000-0002-7263-8228}{ORCID})
}

}
\keyword{internal}
