% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPRAentry-package.R
\docType{package}
\name{qPRAentry-package}
\alias{qPRAentry}
\alias{qPRAentry-package}
\title{qPRAentry: quantitative Pest Risk Assessment entry step calculation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package is meant to support risk assessors in performing the entry step of the quantitative Pest Risk Assessments. It encompasses functions for calculating and redistributing the quantity of commodities imported from countries where a plant pest of interest is present and estimating the amount of a pest entering a risk assessment area (in terms of founder populations landing in each country/area/region). The package comprises two Shiny apps: one for computing the potentially infected/infested quantity of commodities imported by a country of interest from third-countries where a plant pest is present and redistributing this quantity across regions, and the other for executing the pathway model. The package and functions used are based on the work of the European Food Safety Authority (EFSA PLH Panel et al., 2018) \url{https://doi.org/10.2903/J.EFSA.2018.5350}.
}
\author{
\strong{Maintainer}: Martina Cendoya \email{cendoya_marmar@gva.es} (\href{https://orcid.org/0000-0002-0134-7781}{ORCID})

Authors:
\itemize{
  \item Maria Chiara Rosace \email{mariachiara.rosace@unicatt.it} (\href{https://orcid.org/0000-0002-7263-8228}{ORCID})
}

}
\keyword{internal}
